[Gradle] - rename '.audioui' package to '.media.ui'
