+++ media_ui/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="edu.artic.audioui" />
+    package="edu.artic.media.ui" />
