+++ media_ui/src/main/java/edu/artic/media/ui/NarrowAudioPlayerFragment.kt
@@ -1,4 +1,4 @@
-package edu.artic.audioui
+package edu.artic.media.ui
 
 import android.content.ComponentName
 import android.content.Intent
