+++ media_ui/src/main/java/edu/artic/media/ui/NarrowAudioPlayerViewModel.kt
@@ -1,4 +1,4 @@
-package edu.artic.audioui
+package edu.artic.media.ui
 
 import edu.artic.viewmodel.BaseViewModel
 import javax.inject.Inject
